Point controller: move event's logic from TripCtrl to PointCtrl
